Revert "install payload and refactor file structure"
This reverts commit 7ed7f70868a39dfe89c7c97c844da2d8e00332c1.
