+++ package.json
@@ -10,16 +10,10 @@
   },
   "dependencies": {
     "@iconify/react": "^6.0.0",
-    "@payloadcms/db-mongodb": "^3.55.1",
-    "@payloadcms/next": "^3.55.1",
-    "@payloadcms/payload-cloud": "^3.55.1",
-    "@payloadcms/richtext-lexical": "^3.55.1",
     "clsx": "^2.1.1",
-    "graphql": "^16.11.0",
     "lightingcss": "^0.1.0",
     "lightningcss": "^1.30.1",
     "next": "^15.3.2",
-    "payload": "^3.55.1",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "react-icons": "^5.5.0",
