--- src/app/(frontend)/layout.tsx
@@ -1,22 +1,0 @@
-import "../globals.css";
-import NavBar from "./components/Navbar";
-import Footer from "./components/Footer";
-import { Montserrat } from "next/font/google";
-
-const montserrat = Montserrat({
-  subsets: ["latin"],
-  weight: ["100", "400", "700", "900"],
-  variable: "--font-montserrat",
-});
-
-export default function AppLayout({
-  children,
-}: Readonly<{ children: React.ReactNode }>) {
-  return (
-    <div className="flex flex-col min-h-screen">
-      <NavBar />
-      <main className="flex-grow">{children}</main>
-      <Footer />
-    </div>
-  );
-}
