--- src/app/(payload)/admin/importMap.js
@@ -1,1 +1,0 @@
-export const importMap = {}
