--- src/app/(payload)/layout.tsx
@@ -1,5 +1,0 @@
-import type { ReactNode } from "react";
-
-export default function PayloadLayout({ children }: { children: ReactNode }) {
-  return <>{children}</>;
-}
