+++ src/app/components/BannerAnimation.tsx
