+++ src/app/components/ExecCard.tsx
