+++ src/app/components/HomePageBanner.tsx
