+++ src/app/components/HomePageBannerAnimation.tsx
