+++ src/app/components/SponsorInfoPanel.tsx
