+++ src/app/layout.tsx
@@ -1,10 +1,34 @@
 import "./globals.css";
-import type { ReactNode } from "react";
+import NavBar from './components/Navbar';
+import Footer from './components/Footer';
+// import Geist fonts via CDN in <head> instead of using 'geist/font'
+import { Montserrat } from 'next/font/google'
 
-export default function RootLayout({ children }: { children: ReactNode }) {
+const montserrat = Montserrat({
+  subsets: ['latin'],
+  weight: ['100', '400', '700', '900'],
+  variable: '--font-montserrat',
+})
+
+export default function RootLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
   return (
     <html lang="en">
-      <body>{children}</body>
+      <body
+        className={`${montserrat.variable} antialiased bg-[#f1eafb]`}>
+        <header>
+          <NavBar />
+        </header>
+        {children}
+        <div className="flex flex-col">
+          <main className="flex-grow">
+          </main>
+          <Footer />
+        </div>
+      </body>
     </html>
   );
 }
