--- src/app/my-route/route.ts
@@ -1,12 +1,0 @@
-import configPromise from '@payload-config'
-import { getPayload } from 'payload'
-
-export const GET = async (request: Request) => {
-  const payload = await getPayload({
-    config: configPromise,
-  })
-
-  return Response.json({
-    message: 'This is an example of a custom route.',
-  })
-}
